+++ example/NotificationHandler.js
@@ -27,6 +27,11 @@ class NotificationHandler {
     }
   }
 
+  // (optional) Called when the user fails to register for remote notifications. Typically occurs when APNS is having issues, or the device is a simulator. (iOS)
+  onRegistrationError(err) {
+    console.log(err);
+  }
+  
   attachRegister(handler) {
     this._onRegister = handler;
   }
@@ -47,6 +52,9 @@ PushNotification.configure({
 
   // (optional) Called when Action is pressed (Android)
   onAction: handler.onAction.bind(handler),
+
+  // (optional) Called when the user fails to register for remote notifications. Typically occurs when APNS is having issues, or the device is a simulator. (iOS)
+  onRegistrationError: handler.onRegistrationError.bind(handler),
 
   // IOS ONLY (optional): default: all - Permissions to register.
   permissions: {